fix build cylic dep
