+++ sdk/src/network/public/index.ts
@@ -1,3 +1,2 @@
 export * from "./fetcher/";
-export * from "./fetcher/fetcher-impl";
 export * from "./parsing";
